Refactor: Move junit package under connectivity package
The junit package is only used by the connectivity package, so it can be
under /connectivity/internal/ instead of /internal/.
